Parallelize AiEvaluateGoalsSystem
I'm not sure why, but this is significantly faster without thread occupancy by other jobs. (It may be two and a half times slower if other jobs can run beside it).

While switching to IJobChunk and NativeStream is the correct solution, I was curious how far this approach would get me. It got me far enough for now.