Fix radar FOV authoring and compute cone bounding boxes for radar instead of sphere
The FOV fix reduced the aggressiveness of the AI but did not offer much of a performance boost.

The custom bounding boxes for the cone had a respectable improvement for FindPairs.LayerLayerSingle. 3.5 ms -> 2 ms. With safety checks, this optimization is exaggerated (10 ms -> 4 ms).

The cost of building the override AABBs is less total across all threads compared to the savings. So this is not just a latency win but also a throughput win.