More Shader Properties and Systems
I added Speed and IntegratedSpeed properties as well as some initial work for vertex warping in wormholes now that I have ruled out pixel-warping. I'm not actually that upset about it since I am GPU-bound anyways. Though I might need to tesselate the capsule somehow *shudder*.

Right now I am hoping that the Speed and Fade properties will fill in the holes of the gravity warp system, which builds two collision layers and runs a bipartite findpairs. The cross buckets might be especially expensive.

I'm thinking for the shader to dot test the heading and position with the view direction and if the signs are opposite, flip the mesh along the view direction. Then reproject the mesh onto some range of the paraboloid's interior space. That should be discontinuity-free.

Not really testing because wormhole graphics are bugged in HR V2.