Fix scene switching execution order bug in core, as well as gamepad and hud improvements
It turns out that SceneManager.LoadScene does not change scene immediately. Instead, the scene swapping happens in EarlyUpdate.UpdatePreloading, which is a Unity PlayerLoopSystem. Because of that, the remainder of the InitializationSystemGroup was executing before the scene loaded. This is what caused SpawnQueues to not be initialized properly.

With these new changes, I am disabling the rest of the ECS systems for the duration of the frame, so that only the DestroyEntitiesOnSceneChangeSystem does anything during the frame. I could also split LatiosWorldInitializationSystemGroup, but have opted not to for now for simplicity's sake.

One caveat of this approach is that there will be a frame rendered with GameObjects and no ECS systems. This could cause artifacts for scenes relying on systems to initialize state or perform runtime GameObject conversion. The best solution is to black out the frame on Awake.

Note to self: Scene management behavior in the framework needs some serious documentation.