AI refactor and SpawnQueues memory leak fix
It was time to give the AI code a bit of an architectural makeover before adding other modules. I decided I wanted to define modular behavior modules so that certain factions could be hilariously dumb in certain aspects. This new architecture has a much cleaner naming convention and better data flow. I think it is a little faster, but it is hard to tell at this point just because everything is GPU bottlenecked anyways. That's something I still need to figure out. I have switched back to single-threading the wandering logic with RNG, as it no longer depends on the scanning system. The job runs before the expensive single-threaded scanning jobs start.

What is definitely fixed is the AI wandering logic. There were a few bugs littered in the old mess that I caught during the refactor. Now the AI wander a lot more and consequently are more likely to target the player.

Also, the authoring tools for the AI have been reorganized, and now random ranges for personalities are a thing. It just adds another level of variability and uniqueness to the AI.

Now that the AI architecture is all cleaned up, hopefully I will be able to add features incrementally.

Oh, I finally realized that I had forgotten to dispose the extra NativeLists I added to SpawnQueues. That's fixed now!