Unity 2020.1 release and DOTS 0.13 update
This was a fun one to update! The only feature that was broken was the bootstrap code, and the fix for that involved lots of deleted code! Setting up the playerloop is intuitive now.

There were a couple nice additions in Collections including NativeReference and NativeHashSet which found homes in this repo.

The third surprise was wormholes rendering correctly again now with HRv2. Now I can start experimenting with fun stuff like gravity warping. But I think I might save that for a future release, as it is becoming time to release 0.2.0.

The fourth surprise was the massive refactor of ENABLE_UNITY_COLLECTION_CHECKS to [Conditional] methods. This seems to be for compatibility with Burst 1.4. While I have not tested with Burst 1.4 yet, and don't really intend to support it for 0.2.0, I refactored the code in a similar manner.

Finally, builds are working again! I had a bug in BurstPatcher where the updated CodeAnalysis dependency ended up exposing MetadataImportOptions like I wanted.

No measurable performance improvements with this update.