Bug fixes and first mission scene
Core bug: GlobalEntityData was being copied to the global entity. This has now been fixed.

The spawning proc gen authoring was missing IDeclareReferencedPrefabs. Whoops.

There were two mathematical bugs with the AI. The first was that flyTowardsPosition and Translation.Value were equal. This caused desiredHeading to be zero which exploded distanceFactor and made a NaN mess. I don't know why this was happening so frequently in the new test scene, but since this is an actual problem that can happen, I patched it. The second bug was when an AI got near the border, the radius for the next wander point ended up being less than the destination radius. This lead to the AI constantly regenerating new wander points and spazzing out. Now in this edge case, the AI will fly towards the center of the world until a valid radius can be found. I may switch this behavior to an arena clamping behavior in the future.

I created a new shader graph and materials for the ship. This is just a test visual to differentiate factions. I don't intend to keep this forever. However, it looks cool enough for an early release. So if I don't get any new ship designs, I will be using these for the 0.2.0 release.