Game Over Conditions and Menuing
Admittedly, the Mission Complete and Mission Failed screens aren't as clean as they could be, and the transition is rough. But the system is functional. Now I just have to implement Mouse and Keyboard controls as well as GUI for showing them and a quit option from the menu and I may cut that for 0.2.0. Pausing and smooth game-over can come with another update.