Core bug fixes and new test scenes
Had to fix a couple of bugs that were bothering me in Core. Even though they somehow didn't affect the game, they were bothering me nonetheless.

The first bug was that I was not deleting the RequestLoadScene component on scene change. This matters if the component is on a DontDestroyOnSceneChange entity or a WorldGlobalEntity.

The second bug was how I was filtering the ManagedStructStorageReactiveSystemGroup. It worked properly if all the components were added or removed the same frame (like creating and destroying an entity, but broke otherwise. Trying to fix this using a single EntityQuery was a pain though, because Unity's scratch allocator buffer was hardcoded to 1024. So for now I am reading all EntityQueries of all the children systems. This is not exactly amazing as checking 14 systems takes 12 microseconds whereas an empty system takes 1 microsecond according to the profiler. It seems I may need to Burst this check in the future.

I also updated some test scenes as I was showing off the game to family.