Initial mouse and keyboard support
I'm not a mouse and keyboard person. So this will definitely require more tuning. But it works. Right click engages lock-motion mode.

I still can't get the cursor to disappear consistently. I suspect this may be a Unity bug. No idea what is up with that. It locks correctly though.