Add Runtime Profiling Graphic
This code is really really messy, and I need to revisit it at some point and make it resolution-independent as well as more tightly-integrated into the framework.

Making this work in builds was a bit of a hassle. The point where Gfx.WaitForPresent is different in the editor vs a build. In the editor, it happens right after my final ECS system runs. In the build, it happens right when the SRP executes before SRP code is executed. This meant that in the build I had to find a PlayerLoopSystem that executed before the update and then subscribe to the SRP beginFrameRendering to stop the timer.

Overall, I am happy with how the profiler otherwise behaves. I need to properly integrate it in the future, because it is quite useful as a tool. Performance for drawing the image isn't quite what it could be, but that isn't really the show-stopper at the moment.

So I learned that my Skybox is eating up the GPU pretty bad. So I think I might try and pre-render a cubemap. There's also a bug in builds where managed objects don't get destroyed properly that is causing scene switching upon mission completion to die real hard. That requires more investigation.