Bug squashing and Unity bug identification
Lots of bugs squashed. First, a couple of bugs related to spawners being decoupled from the ships has been solved. Turns out spawners were being proc-genned out of bounds.

There was also a bug with global data merging not properly deleting the merged entities.

And then there was an issue where the world would unpause before the DestroyEntitiesOnSceneChangeSystem would trigger, which caused a whole bunch of problems in builds but was never seen in the editor. I made the world pausing hack a little cleaner, which brought light to the real issue.

**SCENEMANAGER EVENTS ARE BROKEN!!!!!!!!**
I tried sceneUnload, sceneLoad, and activeSceneChanged. Every one of them will only ever be called once in a build and never again. In the case of sceneUnload, there will be a large latency before input is accepted on the results screen (like 5 to 10 seconds) and then the scene will change, but the main menu or gameplay will be frozen. For the others, the scene gets stuck on the results screen with input hovering still working.

I'll be working on documentation and hope that Unity can fix this bug in a timely manner. Ugh...