v0.2.0 Release!
So I created a few more missions, including a couple new factions. I'm not 100% sold on the coloring but they seem okay enough.

I added a dev code to switch out of a mission back to the main menu. That way people can try the different missions without necessarily completing them.

I also fixed a HybridRenderer issue in builds where I needed more hybrid renderer vram to fit Mission 6. Unfortunately the editor setting does not transfer to builds, so I set it up in the bootstrap instead.

Performance-wise, I am very happy with missions 1-4. They really show off the size of the simulation that can run with butter-smooth framerates. For mission 5, depending on where I look I can either be CPU-bound or GPU-bound. Framerate tanks a bit, but it remains playable and eventually things calm down enough where the framerate picks up. Mission 6 wrecks the system. It gets single-thread bottlenecked by the 50,000 orbital spawners cross bucket checking against the 300,000 bullets. That's not surprising. It is literally the Optimization Adventures #2 bottleneck in a real-world use case. It will be interesting to see if I can make that mission playable in a few more releases of the framework.