First attempt at EnableCommandBuffer with large sync point reduction improvements
I am not 100% confident this is actually safe. There could be an undetected memory leak. I have also only tested single-threaded usage of the backing UnsafeParallelBlockList.

What I do know is that SpawnShipsEnableSystem is working just as before, except now way faster in Mission 6. Before it was regularly taking 20 ms just because it was iterating through all the spawners without Burst. In the massive enable wave, the system would take over a second to process.

Now during that massive first wave, the system takes 80 ms. 2.6 ms is writing to the EnableCommandBuffer, 0.4 ms is sorting the commands (kinda dumb because the commands were single-threaded, but whatever), 3.9 ms to gather the Entities from the LinkedEntityGroup, and the remaining time is in the single EntityManager.RemoveComponent<Disabled>() call.

At this point, I think it is safe to say that the ball is in Unity's court in terms of optimization.

When the system is not handling a massive wave, it typically runs in about 1 ms, and that is still having a non-zero workload. So I'm pretty happy with the performance.

The next thing to do is refactor EnableCommandBuffer into an EntityOperationCommandBuffer as DisableCommandBuffer and DestroyCommandBuffer can also use this NativeContainer.