InstantiateCommandBuffer is working!
I am amazed that this is working! It feels amazing!

It is definitely faster, but I am not sure by how much. It might even be as much as 2x. I will know for sure once I write the next Optimization Adventure and instantiate each bullet twice, once with ECB and once with ICB. That should give me a 1-to-1 comparison. I will maybe switch the order I create the buffers every other frame.

What I do know is that the largest spike I see for ICB is 11 ms on Mission 6, and it is almost always in the single digits. TimeToLive entity destruction often takes much longer.

I'm probably going to change RadixSort to use a key extraction processor instead of an interface. Then I can expose a library of utility key extraction functions like float and int to uint conversions. That will also allow for other containers to be iterated and expose the RadixSort publicly.

Another bug I observed is that if the amount a bullet moves in a frame is greater than the length minus the distance the shooter moves, the bullet will hit the shooter. I need to fix this.

Also, this is my first time exposing Entities using asmref. It seems to work well. I will probably do more of this moving forward. Although I might make this become an [InternalsVisibleTo] relay to a separate asmdef to make the API internal to Core. I will play that by ear.