Box Colliders weekend 1 checkpoint
I have not really validated any of this code. That will come later.

So apparently git missed a few Conditional fixes, so those are now in this commit.

I reordered all the queries to use a consistent BinASpace rather than AinBSpace.

I need to validate that raycasting an Aabb when the ray starts inside returns false.

For sphere vs box, I had to drag in the point queries file. I have not vetted this code, so a lot of it is commented out.

For sphere vs capsule, I started along the path of using a switch case. But I realized in most cases that I was testing all faces and at least 4 of the edges. So I asked myself if there was a brute force method for determining whether a box edge was valid and I figured out a heuristic. With that, I scrapped the massive switch case and made a brute force approach that relies heavily on simd. It might not be as fast as the theoretical switch case, but it was faster to write, and is a little more manageable to optimize in the future.

During that process, I realized that simdFloat3 is missing quite a bit of potentially useful functionality. I added the few pieces I needed, but I should probably revisit it and fill out the other pieces and write some proper documentation for it. That's something I will do when I don't have the mental energy to tackle box vs box.