Box Colliders Weekend 2 Checkpoint
Wow.
I'm starting to question if this is actually faster than GJK (I think it is still probably faster than GJK + EPA). This borrows from SAT to find the closest features along each axis, but then it brute forces the distance evaluations for each axis. For the edge vs edge scenarios, finding the nearest features cuts the number of checks by a factor of 4.

I started enum-hacking the shuffles. I might need to move this optimization to box vs capsule.

A lot of this code is simd, and unlike GJK, this algorithm can heavily benefit from AVX.

Anyways, the next step is to get this algorithm wired in and start testing it and find out how badly it is broken.