Update packages and install new audio packages
So these are the first external git packages used in the project. I wanted to keep the project independent of external software so that everything can be done in Unity or code editors. USFXR fits perfectly. I grabbed the fork that was listed on OpenUPM.

Music was trickier, as I couldn't find a UPM tracker generator and player. So instead I looked for some text-based music description language and found CSound. I also found CSoundUnity. This seems like a viable solution as people can and do modify csd files directly to create music, but tools for handling music traditionally like CabbageAudio and Blue also exist, which I will be experimenting with.

I would still prefer a tracker tool in Unity, but this is the best I can find and so I am going to go with it.