Audio Weekend 1 Checkpoint
And so this next adventure begins. This is the first piece of the puzzle foreshadowing how this audio system is getting put together.

There's a couple of pieces that are different from the original prototype though. For starters, I am doing the ILD processing on the audio thread rather than the job threads. I might choose to do this differently later depending on performance. But for now I am doing it this way as I won't need to keep a running SVF history.

Second, I am implementing subframes which account for when the audio framerate is faster than the graphical framerate.

Third, I have a boolean to enable or disable warnings.

And fourth, I plan to expose the DSPGraph more directly rather than hide it. These means users will be able to blend the sound effects from ECS with custom audio for unique objects and background tracks.

My next step is to implement the LPF node and the ILD mixdown node. But that may not be for a while as I will be out of town and away from my prototype implementation to use as reference. So instead I will probably do some LSSS UI work.