UI Redesign 1
And this is why I don't do web design. Someone, please fix.

Also, I am committing this on Christmas Eve. Don't worry. I spend time with family. But sometimes they are busy doing things that I would just interfere with, so I do this instead.

I'm not totally convinced I like the dot font, although the way it bleeds in bloom is kinda cool. The main two things are that it is monospace and that capital 'I' has the top and bottom bars which fill space (it looks hideous in monospace).

I also fixed the alignment for the in-game HUD faction readouts.

That means the missing UI pieces are the new shaders for the health and boost bars, and the credits screen.

This is my last commit from my laptop. I'll be back at audio on my desktop this weekend.