Audio Checkpoint 2
Not a weekend checkpoint because I am doing this not on a weekend. But it is a clean checkpoint spot.

This commit just adds two DSP Graph nodes.

The first is the StateVariableFilterNode which comes from the DSPGraph samples. It is modified so that the top-level struct is the node. Also, the NativeArray was replaced with a FixedList for storing state. It is one less resource to worry about.

There are three potential optimizations for StateVariableFilterNode coming from the fact that the current implementation is completely scalar. The first is that there are potentially some 2-wide vector operations. However, I doubt the win will be that large. A second win could come from caching v1 and v2 into arrays and then vectorizing the outputBuffer store operation, which would vectorize 4 / 10 multiplies and 3 / 8 adds/subtracts. The last optimization could be having a 4 or 8-port version of this node which interleaves the data, then vectorizes it, and finally outputs it. I'm not sure how much of a win that would be using shuffles. But on some architectures I could see that being really fast with gather-scatter. We'll have to see if this SVF ever becomes a performance hog.

The second node is a simpler mixing node. And it vectorizes nicely. That's all I have to say about it.