Audio Checkpoint 4
I'm pretty happy with the algorithmic implementation here. Compared to the original, I got rid of the incrementing playhead and I also am no longer risking major precision loss when computing the looping source's start point as I am computing the integer and fractional parts separately.

What I am not as happy about is that Burst is not even trying to vectorize most of the sampling loops. It got the oneshots with the matched rates, but that's it. Oh well. I am not going to worry about it now, as I really just need to get this feature in a complete and shipping state first.

Something I'm not totally sure I like or dislike (but I'm leaning more towards like) is that I'm now parallelizing over channel rather than clip. This means the temporary mega buffer is a lot smaller. This comes at a cost of less parallelization. However, I could theoretically clone the channels and reassign some of the clips to the clones for added parallelism at the cost of more memory.