Audio Checkpoint 5
So add the removal of the tracking components to the list of improvements over the prototype! This allows for real clip setting and possibly play on demand in the future (well technically you can play on demand now, but you cannot defer playback).

I still need to source culling and merging as well as the final mixdown and transfer. I think I have an idea now for how to support layers, by having layer entities with a default layer0 created by the default audio system on the worldGlobalEntity. Probably won't make it to the v0.3 release.

Anyways, that wraps up my holiday break. Back to work in C++ land.