Audio Checkpoint 7
So batching happened. I ended up rewriting it twice to get it to where it is. Originally I was working with ranges, but then decided to simplify things and also reduce the memory pressure that the individual weights induced by only paying for weights that pass culling.

I also re-evaluated how I was handling all the FixedLists and realized that memory constraints were getting out of hand. So I put in the effort to convert the IldProfile to a blob asset and added an authoring workflow for it in the form of a ScriptableObject.

So I think my strategy for managing the shared frame counter is to allocate in a SystemState in ECS and deallocate from a DSPGraph update kernel scheduled on destruction of the SystemState. This synchronizes with the creation and destruction of the graph nodes except the counter is allocated early to handle excessive ECS updates.