Audio Checkpoint 9 - I hear things!
We're on the home stretch!

At this point it is integrating more of LSSS to use audio and cleaning everything up.

There were lots of exceptions, lots of Debug.Log expressions, lots of crashes because logging too much makes Unity unstable, and lots of little bug fixes. Yes, I come up with some crazy architectures and mechanisms and simultaneously write some of the most stupid bugs.

Anyways, lots of stuff to do. The UI code needs an overhaul so that sounds can be properly tied to actions in systems. I need to switch more things over to IJobEntityBatch. I need to clean up the logging of the audio driver. I need to rename some things for consistency, especially API type names. I need to rearrange the game loop so that audio can be placed where it belongs more naturally. And of course I still need to validate looped and spatialized sounds and implement BGMs.

But I am very happy with what has been accomplished here!