API Breakages!
Definitely going to need to write an upgrade guide for this one.

Physics was renamed to Psyshock. Audio was renamed to Myri. Those broke asmdef references and namespace usages.

Global entities are now blackboard entities. This lead to a reset of the authoring settings. Yet another bug in [GenerateAuthoringComponent].

I got rid of a few "Latios" prefixes on types. This mostly affected authoring.

Next up is the conversion bootstrapper and the initialization system group rework. Then I get to clean up LSSS and write docs!