Compatibility Cleanup
Wow! That really cleaned up LatiosInitializationSystemGroup and SuperSystem!

I'm not a fan on how less apparent the system ordering in initialization is. And I'm not sure if the ordering is set up correctly relative to the Unity systems. But LSSS is working and there are far fewer hacks in place.

I'm not sure if my conversion configuration code is good. I will need to test compatibility with some other projects.

I switched to using UnsafeUntypedBlobAssetReference in Collider.

Overall I'm pretty happy with this cleanup!

Next up is the UI overhaul!