Menu Rewrite
A lot of code went into this to migrate the functionality from the MonoBehaviour to the system. Fortunately, a lot of it just worked. The only thing that didn't was the sliders, which I'm guessing are a bug with this version. Oh well, the workaround was trivial anyways.

I discovered I could steal some editor assets for the arrows. They look pretty awesome now! The UI navigation is a little funky though. I might need to manually adjust that. I'm not sure though.

Also I found a bug in the Myri converter for clips. So that is fixed now.

All that's left is gameplay sounds, music, and the new ships and levels!