Miscellaneous Cleanup before Design Collab
I'm up way too late and feeling so good about this! Lol.

So the draw distance control is working beautifully! That will hopefully allow more people to enjoy the game. I may need to tweak the mid-tier distance though. I also fixed the UI so things aren't so tiny.

In doing that, I fixed a bug with runtime cubemap changes in the menu. Now that works and you can see the resolution differences on-screen.

I fixed a bug where InstantiateCommandBuffer.ParallelWriter had default sortKey values. Those are only supposed to be there for the non-ParallelWriter.

And lastly, I made all instantiations of all things optional. Ships can choose to not have bullets to fire, or not generate explosions upon destruction. There's also two new things that can be instantiated: hit effects and gun firing effects. Admittedly I haven't actually tested this code at all other than making sure that when the prefabs are null nothing bad happens. But I'm sure we'll discover issues this upcoming collab weekend.

Lastly, Myri is crazy interesting to hear on some of the older missions. It gets loud. I may have to see if someone on the forums or something has a limiter I can give it.

In terms of performance, it is alright. Updating, culling, and weighting are all really fast. So is the mixing thread. I'm pretty sure I could hammer the mixer thread with way more channels and filters and still be sufficient running in synchronous mode.

The system itself is taking a good chunk of time to run though - close to 300 microseconds. And sampling is already chewing up most of the sync point time running up to 2 milliseconds. I may need to optimize that in a future release.

All in all though, I have what I wanted, which is scalable spatialized audio for nearly free. I hope others will enjoy this too!