Fix cameras
Cameras are no longer entities due to the bug introduced with the recent Windows update. No idea what is going on, but since it was broken for me, it was probably broken for others. So I will wait until another Unity upgrade before removing this, even if a Windows update fixes it.

I also made it so that SubSystem asks about the Bootstrap when it fails to find the LatiosWorld. That will help new users trying to use Myri.