Bug fixes and credits update from design session
Apparently I forgot how parenting works in DOTS. That was such a stupid bug! I get that maybe someone would want to separate parenting from positioning, but at the same time, does that really justify needing to add two components to an entity to parent something?

Overall, the design session lead to a better mouse look, a material for the k-wings, a sound effect for the k-wings that is very satisfying (it sounds different from within the cockpit in a realistic way that I wasn't expecting), and more bugfixes.

I'll be slowly updating the designs throughout the week, but for now I need to get some more documentation done while my mind is weekend fresh.

I'm really cutting this close to a February release, and I really wish I had some help, but such is life I guess.