Reorg, bugfixes, and finally a good Mission 1 in Sector 3
So I reorganized the structure of the project and the prefabs. Materials are now with their objects because that's more intuitive in this kind of project. No more v3 folders. There's a little bit of v2 because of legacy prefabs which overrides use. I will clean those up later maybe.

There was a nasty bug in InstantiateCommandBuffer. That started showing up when I had more than one prefab in the command buffer for a given frame. It was a classic "me" mistake, but was definitely interesting to track down since jobs were reporting recreating the worldBlackboardEntity and then trying to find its Parent component.

Another frame-perfect race condition showed up when I killed a ship the exact frame it fired, which meant that it immediately orphans its firing effects which confused the transform system. This is not common and I'm really glad I caught it and fixed it before release.

Lastly, I'm pretty happy with the starjet firing sound. That took me a while to find the right parameter to get the whining, but once I found it, wow does it bring personality!