Bugfixes from video
So I knew about the RunImmediate loophole for a while. I was going to patch it when the next Entities dropped, as it seemed pretty obscure. However, how long has CompoundCollider scaling been broken for?!

Well it is fixed now. And I also fixed some Editor spam warnings by renaming files to not conflict with internal Unity. Tomorrow I get to make a patch release and fix the Optimization Adventure #4 bad tables.

If I am lucky I will still get the video finished before 48 + 8 Jam, but it is tight.