New Feature: EntityWith
One of the first things to come with the new GameplayToolkit, this little thing adds a little bit of verbosity but drastically simplifies collaboration between team members.

It also adds a little bit of code cleanup when doing reference chasing through an entity hierarchy. Although this only showed up in one spot in LSSS where it is still a little rough to justify.