New Feature: Parallel RNG
Ever since the GDC video on noise-based RNG came out, I knew it was the perfect fit for DOTS. The idea is that by having a noise function which takes multiple inputs, you can generate deterministic non-repetitive random numbers in parallel.

So first, I needed to a mechanism to convert noise results into useful ranges. Most of this logic is borrowed from Unity.Mathematics.Random. And then second, I needed to apply the noise as an parallel RNG API. My solution for this was to have an Rng struct which generates a unique random sequence for any given index. It can then be shuffled using the Update method such that the next time a sequence is obtained for a given index, the resulting random values of that sequence will be different.