n - 1 audio rendering and constant audio frames
Instead of having one shots track multiples of audio subframes, subframes are now frames and the ECS side also handles the striding. This allows audio settings changes on the fly to not distort one shot offsets, leading to a more robust experience.

In addition, there is an issue where if two audio thread updates happen between buffer n's start and buffer n+1's submission, then the first audio frame of one shots will be dropped. This is problematic when the worker threads spend a bunch of time sampling. While one could increase the audio frames per update to 2, some may instead prefer to render all audio sources an audio frame ahead. There is now a queue of buffers and all other sorts of magic to make this option a reality.

Also I learned while debugging that one of the menu sfx is only 3 audio frames long on my system.  That's 1/16th of a second. I know musical notes can be even shorter than that, but it really does show how a tiny amount of sound can have quite the impact.