OnNewScene() callback
One of the biggest gotchas with the framework is that people try to operate on sceneBlackboardEntity in OnCreate(). That is never safe, but sometimes seems safe in the editor due to the scene loading before the world.

There's now a safety check for this, as well as a new dedicated callback to set up sceneBlackboardEntity whenever a new sceneBlackboardEntity is created (including in the Editor in play mode). This also cleans up the code and removes the need for first frame sync points.

At this rate, the 0.4 release might happen by the end of the month yet! We'll see if a new DOTS release decides to interrupt that or not.