Updates from internal projects prior to release
Yup. 0.4 is releasing real soon!

A lot of the new stuff was developed against non-public projects, hence a few things in this commit like the new Psyshock stuff and bug fixes seem to come out of nowhere.

I did make a last-minute update to the Rng though. Thanks to a helpful soul in the comments of the GDC video, I learned that on Twitter there was a new version of the noise function. Even better, it came with a proper license! I can sleep better at night now.

Builds are primed. Docs are written. I am working on the Optimization Adventure as I type this, and once that is done, I just have to get the final touches in place and publish. This is by far the most relaxed I have been for a feature release. Feels nice!