Update Unity to 2020.3.30f1 LTS
So turns out there were a few breaks. One of the issues happened to be something another user encountered long before I did related to the sky renderer. Surprisingly that was the only thing broken with the sky, so the sky API has been fairly stable.

The second thing was the exposure. Apparently this was bugged in 2020.1 (which I kinda knew), but there was a charm to that look and now everything turned black. Fortunately modifying the sky profile to add exposure compensation restored the original look!

The last thing is that in this temporary state, deferred arrays from lists are broken. Surprisingly, this only affects Myri. For now I just disabled Myri. Hopefully I get to restore it after updating to new DOTS.