Working in DOTS 0.50 - sort of...
There's a few hacks here and there to get cleaned up. A noticeable surprise was EET being broken, but just as surprising, using a raw EntityManager worked in its place (this was in a Job.Run() inside InstantiateCommandBuffer playback).

Another surprise was the new caching also forces change filter checks onto the main thread. So that code got replaced with IJobEntityBatch.

Burst isn't working when fully when safety checks are enabled. That's frustrating.

But Myri just worked, so that's fun.

Anyways, still much more to explore and cleanup. I think full compatibility should be possible once the Burst issue is resolved.