Framework 0.4.5 - THE ISystem INVASION
I am really happy with how much of the framework I got working in ISystem. What is nice is that I don't even have to create an alternate interface or anything like that. I can just extend SystemState to get most of the properties in place.

I don't actually see much of an improvement with ISystem performance-wise, but that's because LSSS has bottlenecks in other places currently.

Also that Burst issue is resolved! So while there may be more optimizations and other improvements, I think it is time to ship this patch. Now onto the Changelogs (I'm hyped)!