Begin 0.5 Migration
Here's to finally moving the non-Kinemation 0.5 work to a public repo. This is the first time people will see the convex colliders and triangle colliders, as well as chunk component support for Fluent, and the collider debug drawers.

I've also been making small tweaks and improvements where quick wins present themselves. I increased the add tag limit on InstantiateCommandBuffer (I still need to add tags to the no-initialize case).

I also tried migrating Fluent to EntityQueryDescBuilder, but that hard-crashes. :(

Still plenty of small things to do, like add Burst branch intrinsics, formalize SIFE, add a few more InstantiateCommandBuffer versions (the mini demos want 7).

The next big feature, smart blobbers, will also be developed in this repo, as that is a prerequisite for allowing users to convert their animation clips and store the results on arbitrary components. But it may also serve useful for Psyshock and Myri, which is why this is a good repo to test.

And I should probably make a new faction to show off Convex Colliders...
So much to do!