Audio Deduplication with Smart Blobbers
It is working. In my basic tests, I have both shared clip deduplication and identical clip deduplication. There's a couple of nuances with the API I might still need to work through. It still takes a lot of lines of code to implement. But there's a definite decrease in complexity and increase in features.

Listeners and compound colliders will both likely use the context variant. For convex meshes, I need to implement the hashing variant as well as the mesh hashing service.