Listener Profiles with Smart Blobbers
Got to try out the simple API with this one by using a little hack to pre-convert the blob and assign it to the converter. I don't think listeners will be impacted too badly by lack of source asset deduplication and such.

Psyshock will be next. That one will be harder because I have API challenges for transferring arrays of colliders to the smart blobber for compounds, as well as all the mesh stuff for convex colliders.