Update to 0.5.1 and begin 0.6.0
Yeah. 0.5.1 is out now. LSSS was not a major test project during its development. But it still is useful for validation.

But now it is time to start investigating 0.6.0, and LSSS is more useful. I did say on the forums I was going to take a break from new features. But I got bored today. So I am doing new feature dev.