Big 0.6 Overhaul
This was a lot of work that got all wrapped together.

The first big change was the FindPairs optimization investigation. The conclusion was that negating the max values of the y and z components to remove the need of swizzling every pair was a safe and effective optimization. While the code is left around, unrolling had inconsistent behavior, mostly positive for LayerSelf, but negative for LayerLayer. Sentinels may be a future option, but they can cause a significant burden to managing buckets. Something to still think about.

But the bigger optimization was reworking FindPairsResult to be passed in as a reference and to make all of its fields properties that lazy load from the layer. It is a breaking change, but in the end I think it is worth it.

The second big change was the addition of layer queries in the form of FindObjects and immediate shorthands for Raycast, DistanceBetween, and ColliderCast. This code isn't really tested as while it is used in LSSS for ship radar line-of-sight, there are no walls in any of the missions to actually query against. However, layer building seems to work correctly. I at least don't get bounds errors. That's relevant because I am now building a 1D interval tree. The interval tree is used as a sort of inside-out test. It finds AABBs that the query AABB's x value starts inside of, and does so in O(log n) time. But even better, it is built in O(n) time given sorted xmin values. I probably still need to optimize the traversal at the micro-architecture layer, but the tree at least seems like a good complement for the SAP without building a completely independent solution like a BVH.

The rewrite of ship radar scanning had a lot of impacts. First, it brought the need to do intermediate cross bucket caching. But even that wasn't enough because the processors were heavy and still creating bubbles. [NativeDisableContainerSafetyRestriction] didn't work correctly with CDFE, so instead I made AiShipRadarScanResults arrays for each faction and then wrote them back at the end. This made job scheduling tight. So tight, that lots of small jobs that used to squeeze in, didn't, causing a little bit of an airy mess in the profiler at the tail end, with the audio jobs starting sooner and overlapping. This part ended up being a slight regression. Another issue is that there are less JobHandle.CombineDependency() calls in the new version and the system does a lot of scheduling work. This results in the worker threads being briefly starved on otherwise well-running missions.

After that, I figured out how to make custom containers work with custom allocators, so I reworked all the containers as well as CollisionLayer. Now syncPoint uses UpdateAllocator, which should be faster.

Lastly, I migrated a bunch of systems to ISystem. The performance benefit is measurable. I did not do type handle caching though.

I fixed some minor bugs in the process and cleaned up a few things. Next up is to merge this with the Kinemation optimizations and improvements. Though, that will be a weeknight thing.